// my-app/src/app/services/llamada/llamada-interface.ts
export interface LlamadaInterface {
    id?: number;
    note?: string;
    day?: any;
    contact_id?: number;
    name?: string;
    surname?: string;
    institution ?: string;
    user_id?: number;
    user_name?: string;
    office_id?: number;
    created_at?: string;
    updated_at?: string;
}